+++ Runtime/Screens/BaseScreen.cs
@@ -7,9 +7,8 @@ using System;
 namespace RedHeadToolz.Screens
 {
     // A class that must be extended from if you expect to use it with the ScreenManager
-    public class BaseScreen : MonoBehaviour
+    public abstract class BaseScreen : MonoBehaviour
     {
-        [SerializeField] protected Transform _root;
         [SerializeField] protected GameObject _inputLock;
         [SerializeField] private bool _hideStack = false;
         public bool HideStack {
@@ -34,50 +33,11 @@ namespace RedHeadToolz.Screens
             }
         }
 
-        public virtual void Show(Action callback = null)
-        {
-            _lockInput = true;
-            _root.localScale = Vector3.zero;
-            
-            Sequence showSequence = DOTween.Sequence();
-            
-            showSequence.Append(_root.DOScale(1.1f, 0.4f).SetEase(Ease.OutSine));
-            showSequence.Append(_root.DOScale(1.0f, 0.1f)
-                .SetEase(Ease.OutSine)
-                .OnComplete(()=>{
-                    if(callback != null) callback();
-                    _lockInput = false;
-                }));
+        public abstract void Show(Action callback = null);
 
-            _showing = true;
-            showSequence.Play();
-        }
+        public abstract void ShowImmediate();
 
-        public virtual void ShowImmediate()
-        {
-            _showing = true;
-            _root.localScale = Vector3.one;
-            _lockInput = false;
-        }
-
-        public virtual void Hide(Action callback = null)
-        {
-            _lockInput = true;
-            _root.localScale = Vector3.one;
-            
-            Sequence showSequence = DOTween.Sequence();
-            
-            showSequence.Append(_root.DOScale(1.1f, 0.1f).SetEase(Ease.OutSine));
-            showSequence.Append(
-                _root.DOScale(0f, 0.4f)
-                .SetEase(Ease.OutSine)
-                .OnComplete(()=>{
-                    if(callback != null) callback();
-                }));
-            
-            _showing = false;
-            showSequence.Play();
-        }
+        public abstract void Hide(Action callback = null);
 
         public virtual void Close(Action callback = null)
         {