+++ Editor/IncrementVersion.cs
@@ -6,6 +6,20 @@ using System;
 
 public static class IncrementVersion
 {
+    [MenuItem("RedHeadToolz/Version/Current Version: 0.0.0", false, 0)]
+    public static void ShowCurrentVersion()
+    {
+        // This method is just a placeholder to show the current version in the menu
+    }
+
+    [MenuItem("RedHeadToolz/Version/Current Version: 0.0.0", true)]
+    public static bool ShowCurrentVersionValidation()
+    {
+        Menu.SetChecked("RedHeadToolz/Version/Current Version: 0.0.0", false);
+        Menu.SetChecked($"RedHeadToolz/Version/Current Version: {PlayerSettings.bundleVersion}", true);
+        return false; // Return false to disable the menu item
+    }
+
     [MenuItem("RedHeadToolz/Version/Increment Major")]
     public static void IncrementMajor()
     {