+++ Runtime/Initialization/InitController.cs
@@ -7,6 +7,8 @@ namespace RedHeadToolz.Tools
 {
     public class InitController : MonoBehaviour
     {
+        [SerializeField] GeneralManager _gm;
+        [SerializeField] private string _sceneToLoad = "MainMenu";
         void Start()
         {
             Screen.sleepTimeout = SleepTimeout.NeverSleep;
@@ -16,11 +18,14 @@ namespace RedHeadToolz.Tools
 
         void Update()
         {
+            // move thesse two into managers
             if (AudioController.Instance == null) return;
 
             if(ScreenManager.Instance == null) return;
+            
+            if(_gm.IsInitialized == false) return;
 
-            SceneManager.LoadScene("MainMenu");
+            SceneManager.LoadScene(_sceneToLoad);
         }
     }
 }