+++ Runtime/Screens/ScreenManager.cs
@@ -8,6 +8,8 @@ namespace RedHeadToolz.Screens
 {
     public class ScreenManager : Singleton<ScreenManager>
     {
+        // move this onto a prefab with a cavas on the same object
+        // may also need to spawn in my own eventsystem
         [SerializeField] private List<BaseScreen> ScreenList;
         private List<BaseScreen> _screenStack = new List<BaseScreen>();