+++ Runtime/Screens/FadeScreen.cs
@@ -1,7 +1,6 @@
 using UnityEngine;
 using DG.Tweening;
 using System;
-using RedHeadToolz.Debugging;
 
 namespace RedHeadToolz.Screens
 {