+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.5]
+
+### Added
+Closed Captions System
+
 ## [0.1.4] 
 
 ### Added